refactor: extract orb slam bootstrap flow
